Add files via upload
im going ahead an uploading this so i don't get docked for late. I added a few codes to my html and css files for portfolio and contact pages and im having trouble figuring out the position for the /class ".main"/ on those two pages but my code ran like i wanted it to for the about me page.

im also having trouble keeping the ".footer" at the bottom of both portfolio and contact page. so im going to keep working on this till i figure it out. for now this is what i have.